Move the ShortenedPath function to a static helpers class in the EaPdf project; fix some newly discovered problems with rendering HTML and CSS; remove a couple options from the config and command line to simplify the first release; add more validations to the command line args and config values;

+++ TestEaPdf/TestCommandLine.cs
@@ -54,7 +54,8 @@ namespace UIUCLibrary.TestEaPdf
             var inFilePath = args[1];
             var outFilePath = args[3];
 
-            Directory.Delete(outFilePath, true);
+            if(Directory.Exists(outFilePath))
+                Directory.Delete(outFilePath, true);
 
             var xmlFile = new FileInfo(FilePathHelpers.GetXmlOutputFilePath(Path.GetFullPath(outFilePath), Path.GetFullPath(inFilePath)));
             var pdfFile = new FileInfo(Path.ChangeExtension(xmlFile.FullName, ".pdf"));
